% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_truncate}
\alias{g_truncate}
\title{Helper function to truncate values away from 0 and 1}
\usage{
g_truncate(g, tol, bound_away_from = c(0, 1))
}
\arguments{
\item{tol}{Truncation tolerance}

\item{bound_away_from}{A vector of 0 and/or 1.}

\item{x}{Vector of values to truncate}
}
\value{
Vector with truncated values replaced by \code{tol}.
}
\description{
Helper function to truncate values away from 0 and 1
}
