% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ci}
\alias{get_ci}
\title{Helper function to compute 95 percent confidence intervals on log-scale}
\usage{
get_ci(est, grad, cov_matrix)
}
\arguments{
\item{est}{The effect estimate (on original scale)}

\item{grad}{The gradient of the effect transformation}

\item{cov_matrix}{The estimated covariance matrix of the counterfactual means}
}
\value{
A vector of confidence interval limits.
}
\description{
Helper function to compute 95 percent confidence intervals on log-scale
}
