% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_D1a}
\alias{make_D1a}
\title{Helper function make outcome of extra nuisance regression}
\usage{
make_D1a(a, A, Y, C, gA, gC, gAS, QY_WAS)
}
\description{
Helper function make outcome of extra nuisance regression
}
