% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif.R
\name{make_eif_ya1_sa2}
\alias{make_eif_ya1_sa2}
\title{Evaluate EIF of E[Y(a1, S(a2))] on data}
\usage{
make_eif_ya1_sa2(
  a1,
  a2,
  R,
  gR,
  A,
  C,
  gA,
  gC,
  gAS,
  Y,
  QY_WAS,
  QD_WACY,
  QY_WACY,
  QY_W
)
}
\arguments{
\item{a1}{Either 0 or 1}

\item{a2}{Should be \code{1 - a1}}

\item{R}{A numeric indicator of having mediator measured}

\item{gR}{Estimate of P(R = 1 | W_i, A_i, C_i, C_iY_i)}

\item{A}{A numeric vector of binary treatment assignment}

\item{C}{A numeric indicator of having the outcome \code{Y} measured, 
i.e., a measure of being not right-censored.}

\item{gA}{Estimate of P(A = a1 | W_i), i = 1,...,n}

\item{gC}{Estimate of P(C = 1 | A = a1, W_i)}

\item{gAS}{Estimate of P(A = a1 | S_i, W_i)}

\item{QY_WAS}{Estimate of E[Y | A = a1, W_i, S_i]}

\item{QD_WACY}{Estimate of extra nuisance regression}

\item{QY_WACY}{Estimate of E[QY_WAS | A = a2, W, C, CY]}

\item{QY_W}{Estimate of E[QY_WACY | A = a2, W]}
}
\value{
An n-length vector of the estimated EIF evaluated on the observations
}
\description{
Evaluate EIF of E[Y(a1, S(a2))] on data
}
