% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif.R
\name{make_eif_ya_sa}
\alias{make_eif_ya_sa}
\title{evaluate EIF of E[Y(a, S(a))] on data}
\usage{
make_eif_ya_sa(a, A, C, gA, gC, Y, QY_WA)
}
\arguments{
\item{a}{value a}

\item{gA}{P(A = a | ...)}

\item{gC}{P(C = 1 | A = a, ...)}

\item{QY_WA}{E[Y | A = a, W]}
}
\description{
evaluate EIF of E[Y(a, S(a))] on data
}
