% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmed2.R
\name{natmed2}
\alias{natmed2}
\title{Natural in/direct effects with two-phase sampling of the mediator}
\usage{
natmed2(
  W,
  A,
  R,
  S,
  C,
  Y,
  glm_gR = ".^2",
  gRn = NULL,
  glm_gC = ".",
  SL_gC = NULL,
  glm_gA = ".",
  glm_gAS = paste0(paste0(colnames(W), collapse = " + "), " + S"),
  SL_gAS,
  glm_QY_WAS = paste0(paste0(colnames(W), collapse = " + "), "+ A + S"),
  SL_QY_WAS,
  glm_QY_WACY = ".",
  SL_QY_WACY,
  glm_QD_WACY = ".",
  SL_QD_WACY,
  glm_QD_WACY_lazy = ".",
  SL_QD_WACY_lazy,
  glm_QY_W = ".",
  SL_QY_W,
  glm_QY_WA = ".",
  SL_QY_WA,
  lazy = FALSE,
  seed = 1,
  tol_gA = 1/sqrt(length(A)),
  tol_gAS = 1/sqrt(sum(R == 1)),
  tol_gC = 1/sqrt(length(A)),
  ...
)
}
\arguments{
\item{W}{A \code{data.frame} of baseline covariates}

\item{A}{A numeric vector of binary treatment assignment}

\item{R}{A numeric indicator of having mediator measured}

\item{S}{A numeric value of the mediator}

\item{C}{A numeric indicator of having the outcome \code{Y} measured, 
i.e., a measure of being not right-censored.}

\item{Y}{The binary outcome of interest}

\item{glm_gR}{A glm formula for the RHS of the estimation of 
P(R = 1 | W, A, C, CY). The variable-names that can be used in the formula are:
\code{colnames(W)}, \code{A}, \code{CY11} (indicator that \code{C}=1 and \code{Y}=1),
and \code{CY10} (indicator that \code{C}=1 and \code{Y}=0).}

\item{gRn}{A vector of known or external estimate of P(R = 1 | W, A, C, CY). If supplied then
\code{glm_gR} is ignored.}

\item{glm_gC}{A glm formula for the RHS of the estimation of 
P(C = 1 | A, W). Defaults to a main terms regression.}

\item{SL_gC}{A \code{SuperLearner} library for estimation of censoring probability
described above.}

\item{glm_gAS}{A glm formula for the RHS of the estimation of P(A = 1 | W, S). Defaults
to a main terms regression.}

\item{SL_gAS}{A \code{SuperLearner} library for estimation of the mediator-conditional
propensity score described above.}

\item{glm_QY_WAS}{A glm formula for the RHS of the estimation of P(Y = 1 | A, W, S). Defaults
to a main terms regression. This regression is fit using inverse weights \code{R}/\code{gRn}, which may trigger
warnings from \code{glm} about the outcome not being between 0 and 1. These warnings are 
safe to ignore.}

\item{SL_QY_WAS}{A \code{SuperLearner} library for estimation of the outcome regression 
described above. This regression is fit using inverse weights \code{R}/\code{gRn}.}

\item{glm_QY_WACY}{A glm formula for the RHS of the estimation of E[P(Y = 1 | W, A = a1, S) | W, A = a2, C, CY].
The available names for the regression formula are: \code{colnames(W)}, \code{CY11}, \code{CY10}.}

\item{SL_QY_WACY}{A \code{SuperLearner} library for estimation of the sequential outcome regression
described above.}

\item{glm_QD_WACY}{A glm formula for the RHS of the estimation of 
E[I(A = a1, C = 1) / (gAn gCn) gAn / (1 - gAN) (1 - gAS) / gAS (Y - QY_WAS) | R = 1, W, A = a1, C, CY]. 
The available names for the regression formula are: \code{colnames(W)}, \code{CY11}, \code{CY10}.}

\item{SL_QD_WACY}{A \code{SuperLearner} library for estimation of a relevant piece of the projection
of the full data EIF onto the tangent space of gR.}

\item{glm_QY_W}{A \code{glm} formula for the RHS of the estimation of 
E[QY_WACY | A = a2, W].}

\item{SL_QY_W}{A \code{SuperLearner} library for estimation of the final sequential regression.}

\item{glm_QY_WA}{A \code{glm} formula for the outcome regression E[Y | A, W], which is used
in estimation of E[Y(a, S(a))]. Available variable names are: \code{colnames(W)}, \code{A}.}

\item{SL_QY_WA}{A \code{SuperLearner} library for estimation of the outcome regression used to 
compute estimates of E[Y(a, S(a))].}

\item{lazy}{If \code{TRUE}, then the non-simplified version of the EIF is used, in lieu of
the extra sequential regression.}

\item{seed}{The seed set before each \code{SuperLearner} fit.}

\item{...}{Other options (not currently used)}
}
\value{
A list with named entries \describe{
\item{risk}{A \code{data.frame} that contains point estimates and confidence intervals 
of several counterfactual probabilities. Two confidence intervals are shown. \code{cil} 
and \code{ciu} indicate the limits of a 95% confidence interval based on the empirical 
variance of the estimated influence function. \code{cil_cv} and \code{ciu_cv} indicate
the limits of a 95% confidence interval based on a cross-validated estimate of the variance 
of the estimated influence function. If \code{SuperLearner} was not used for any of the 
nuisance regressions, then the two confidence intervals will coincide. }
\item{eff}{A \code{data.frame} that shows the total, direct, and indirect effects along 
with confidence intervals. These intervals are created on the log-scale using the delta-method.}
\item{eff2}{A \code{data.frame} that shows the total, direct, and indirect effects based
on an alternative decomposition.}
\item{cov}{The estimated covariance matrix of the estimates}
\item{cov_cv}{The cross-validated estimate of the covariance matrix of the estimates}
}
}
\description{
Natural in/direct effects with two-phase sampling of the mediator
}
\examples{
n <- 500
W1 <- rbinom(n, 1, 0.5)
W2 <- rnorm(n, 0, 1)
A <- rbinom(n, 1, 0.5)
S <- W1 / 4 - W2 / 3 + A + rnorm(n)
Y <- rbinom(n, 1, plogis(-2 + A + W1 / 2 - S / 2))

# add censoring
C <- rbinom(n, 1, plogis(2 + W1 / 2 - W2 / 3))
# arbitrary fill in
Y[C == 0] <- -999

R <- rep(0, n)
# case-cohort sampling
R <- rbinom(n, 1, 0.25)
R[Y == 1] <- 1

fit <- natmed2(
  W = data.frame(W1 = W1, W2 = W2), 
  A = A, R = R, S = S, C = C, Y = Y
)
}
