% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natmed2_fx.R
\name{natmed2_fx}
\alias{natmed2_fx}
\title{Natural in/direct effects with two-phase sampling of the mediator}
\usage{
natmed2_fx(
  W,
  B,
  A,
  T,
  R,
  S,
  Y,
  glm_gR = ".^2",
  gRn = NULL,
  glm_gA = ".",
  SL_gA,
  glm_gAS = paste0(paste0(colnames(W), collapse = " + "), " + ", paste0(colnames(S),
    collapse = " + ")),
  SL_gAS,
  glm_QY_WAS = paste0(paste0(colnames(W), collapse = " + "), " + A + ",
    paste0(colnames(S), collapse = " + ")),
  SL_QY_WAS,
  glm_QD = ".",
  SL_QD,
  glm_QY_W = ".",
  SL_QY_W,
  glm_QY_WA = ".",
  SL_QY_WA,
  seed = 1,
  tol_gA = 1/sqrt(length(A)),
  tol_gAS = 1/sqrt(sum(R == 1)),
  ...
)
}
\arguments{
\item{W}{A \code{data.frame} of baseline covariates}

\item{B}{A \code{vector} of 0/1 indicating whether obs. is in bridging population}

\item{A}{A numeric vector of treatment assignment}

\item{T}{A numeric vector of trial}

\item{R}{A numeric indicator of having mediator measured in phase 3 study}

\item{S}{A \code{data.frame} of measured mediators}

\item{Y}{The binary outcome of interest}

\item{glm_gR}{A glm formula for the RHS of the estimation of 
P(R = 1 | W, A, T, Y). The variable-names that can be used in the formula are:
\code{colnames(W)}, \code{A}, \code{T}, and \code{Y}.}

\item{gRn}{A vector of known or external estimate of P(R = 1 | W, A, T, Y). If supplied then
\code{glm_gR} is ignored.}

\item{glm_gA}{A glm formula for the RHS of the estimation of P(A = 1 | W). Defaults
to a main terms regression.}

\item{SL_gA}{A \code{SuperLearner} library for estimation of P(A = 1 | W).}

\item{glm_gAS}{A glm formula for the RHS of the estimation of P(A = 1 | W, S)
and P(A = 2 | A != 1, W, S). Defaults to a main terms regression.}

\item{SL_gAS}{A \code{SuperLearner} library for estimation of P(A = 1 | W, S)
and P(A = 2 | A != 1, W, S).}

\item{glm_QY_WAS}{A glm formula for the RHS of the estimation of P(Y = 1 | T = 2, A, W, S). Defaults
to a main terms regression. This regression is fit using inverse weights \code{R}/\code{gRn}, which may trigger
warnings from \code{glm} about the outcome not being between 0 and 1. These warnings are 
safe to ignore.}

\item{SL_QY_WAS}{A \code{SuperLearner} library for estimation of the outcome regression 
described above. This regression is fit using inverse weights \code{R}/\code{gRn}.}

\item{glm_QD}{A glm formula for the RHS of the estimation of 
E[( 1 / (gA2n) ) ( gA2n / (gA1n) ) ( (gA1S) / gA2S ) (Y - QY_WAS) | T = 2, A = 2, W, Y]. 
The available names for the regression formula are: \code{colnames(W)}, \code{Y}.}

\item{SL_QD}{A \code{SuperLearner} library for estimation of a relevant piece of the projection
of the full data EIF onto the tangent space of gR.}

\item{glm_QY_W}{A \code{glm} formula for the RHS of the estimation of 
E[QY_WAS | T = 1, A = 1, W].}

\item{SL_QY_W}{A \code{SuperLearner} library for estimation of 
E[QY_WAS | T = 1, A = 1, W].}

\item{glm_QY_WA}{A \code{glm} formula for the outcome regression E[Y | A, W], which is used
in estimation of E[Y(a, S(a))]. Available variable names are: \code{colnames(W)}, \code{A}.}

\item{SL_QY_WA}{A \code{SuperLearner} library for estimation of the outcome regression used to 
compute estimates of E[Y(a, S(a))].}

\item{seed}{The seed set before each \code{SuperLearner} fit.}

\item{...}{Other options (not currently used)}
}
\value{
A list with named entries \describe{
\item{blah}{blah blah}
}
}
\description{
Natural in/direct effects with two-phase sampling of the mediator
}
\examples{
n <- 500
W1 <- rbinom(n, 1, 0.5)
W2 <- rnorm(n, 0, 1)
A <- rbinom(n, 1, 0.5)
S <- W1 / 4 - W2 / 3 + A + rnorm(n)
Y <- rbinom(n, 1, plogis(-2 + A + W1 / 2 - S / 2))

# add censoring
C <- rbinom(n, 1, plogis(2 + W1 / 2 - W2 / 3))
# arbitrary fill in
Y[C == 0] <- -999

R <- rep(0, n)
# case-cohort sampling
R <- rbinom(n, 1, 0.25)
R[Y == 1] <- 1

fit <- natmed2(
  W = data.frame(W1 = W1, W2 = W2), 
  A = A, R = R, S = S, C = C, Y = Y
)
}
