% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_nas}
\alias{replace_nas}
\title{Helper function to replace missing values}
\usage{
replace_nas(x, val = -999)
}
\arguments{
\item{x}{The vector with missing values to be replaced}

\item{val}{The value to replace with}
}
\value{
A vector of the same length as \code{x}, but with \code{NA}s replaced.
}
\description{
Helper function to replace missing values
}
